+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/ExampleData.java
@@ -50,7 +50,8 @@ public class ExampleData {
                 "Can't protect file",
                 "View file without collaboration",
                 "Can’t submit forms",
-                "Can't refresh outdated file"
+                "Can't refresh outdated file",
+                "Tour of tips when opening a document"
         );
 
         // the description for user 1