+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- java-spring: features tips for anonymous
 - ruby: features tips for anonymous
 - python: features tips for anonymous
 - php-laravel: features tips for anonymous